% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forward-backward.R
\name{en_masse_to_df}
\alias{en_masse_to_df}
\title{Convert from class en.masse to class data.frame}
\usage{
en_masse_to_df(class.en.masse)
}
\arguments{
\item{class.en.masse}{a list of class en.masse}
}
\value{
a data frame with 6 columns, c("Tetrad", "Spore", "Chr", "Snp", "p0", "p1")
}
\description{
Converts from simlated data of class en.masse to class data.frame
}

