% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forward-backward.R
\name{est_fwd_back}
\alias{est_fwd_back}
\title{Needs updating: Inferring genotypic states using the Forward-Backward algorithm}
\usage{
est_fwd_back(snp.dat, p.assign, scale)
}
\arguments{
\item{snp.dat}{an object of class \code{snp.recom} that gives the bi-allelic read counts
(e.g., from bowtie2 or from \code{\link{simulate_coverage}}) of each parental type
for each spore along a given chromosome}

\item{p.assign}{a numeric between 0 and 1 (exclusive) that specifies the probability of
correct sequencing assignment (i.e., 1 - sequencing error rate) plus the probability of
new a new mutation at a given snp (see details)}

\item{tetrad.id}{(optional) a character or numeric giving the tetrad identity.
If snp.dat is of class \code{forward.backward} then \code{tetrad.id} will be inherited
from that class.}

\item{chr.name}{(optional) a character (typical) or numeric giving the chromosome identity.
If snp.dat is of class \code{forward.backward} then \code{chr.name} will be inherited
from that class.}

\item{p.trans}{the per snp (check that) recombination (i.e., transition) probability.}
}
\value{
A list of class \code{forward.backward} containing 4 elements corresponding to each
of the four tetrads. Each element is itself a list containing:
\itemize{
    \item The input snp.data
    \item spore_number an integer specifying the spore number
    \item tetrad.id an integer specifying the tetrad number
    \item chr.name a numeric or character specifying the name of the chromosome
    \item snp.locations a vector specifying the snp locations along the chromosome
    \item p.assign a numeric between 0 and 1 specifying the probability of correct
    sequence assignment
    \item p.trans a numeric between 0 and 1 specifying the per base pair recombination
    rate (transition probability)
    \item emissions a matrix of 2 columns by length(snp.locations) rows specifying
    the emission probabilities (see details)
    \item forward a matrix of 2 columns by length(snp.locations) rows giving the scaled
    forward probabilities
    \item backward a matrix of 2 columns by length(snp.locations) rows giving the scaled
    backward probabilities
    \item scale a vector of length snp.locations giving the forward scaling factors
    \item posterior a matrix of 2 columns by length(snp.locations) rows giving the posterior
    probabilities for each parental state
    \item states_inferred a vector of length snp.locations giving the state with the
    highest posterior. In the event of a tie, a state is picked randomly.
    \item lnL a numeric giving the total likelihood of the data (see details).
}
}
\description{
Uses the forward-backward algorithm to estimate ancestral genotypes
along a given chromosome for a given genotyped tetrad or simulated data.
}
\details{
\code{tetrad_estimate_anc_fwd_back} attempts to estimate
parental genotypic 'states' along a chromosome given empirical or
simulated F2 cross data were all four spores of a tetrad are genotyped.
Next-generation data inherits both sequencing error and missing data
-- especially when sequencing coverage is low. In these two cases
(sequence error and missing data) the parental state is ambiguous
or unknown, respectively. \code{tetrad_estimate_anc_fwd_back} takes
these uncertainties into account.

The three steps taken in \code{tetrad_estimate_anc_fwd_back} are:
\enumerate{
 \item Calculate forward probabilities for each state (5' to 3')
 \item Calculate backward probabilities for each state (3' to 5')
 \item From these two probabilities, calculate the posterior probability
 that a snp location is of a given parental state.
}

The two element vector of forward probabilities for each parental state
\equ{f_{i}} at each position, i, are calculated as:
\dequ{f_{i} = e_{i}T_{i}f_{i-1}}

where \equ{e_{i}} is the emission probabilities for each state (see below), \equ{T_{i}}
is a 2-by-2 matrix describing the transition (recombination) probabilities
between two states, and \equ{f_{i-1}} is the forward probability at the
previous position along the chromosome (5' of position i). It is assumed
that each state is equally likely to occur at the first position.


The emmission probabilities are calculated independently for
each snp and depends on the sequence reads assigned to parent "0"
and parent "1" and \code{p.assign}. The emission probabilities are
calculated using the binomial equation. For example, the emission
probability for parental state "0" at snp position i is:

\dequ{e_{i,0} = {n \choose{k0}} p^{k0} (1-p)^{n-k0}}

where n is the sum of the number of reads from both parents and p
is \code{p.assign}.

Like the emission probabilities, the transition probabilities are also
calculated for each snp position. This accounts for the displacement
between snps.  For example if the per base recombination rate is
0.001 and two snps are 10 base pairs apart, the transition probability of
is 0.01 (i.e., the probability of not recombining would be 0.99).

To avoid underflow, we rescale the forward (and backward) probabilities
each iteration to that they sum to unity.

The backward probabilities are calculated similarly but in the 3' to 5'
direction. It is assumed that the backward probability is equally likely
in each state. Following Durbin et al (1998) the backward probability at
snp position i is:

\dequ{b_{i} = T_{i}e_{i+1}b_{i+1}}

Again, these probabilities are rescale to avoid underflow.

The posterior probability that the state at position i is k, \equ{\pi_{i}=k}
given the observed sequence read counts for each parent at position i, \equ{x_{i}}
is calculated by:

\dequ{P(\pi_{i} = k | x_{i}) = \frac{f_{i} b{i}}{\sum\limits_{k=0}^{1} f_{i}^{(k)} b{i}^{(k)}}}

where the super scripts in the denominator are vector indices.

Finally, we can determine the log likelihood of the whole sequence
of observations by summing up the log of all the scale factors in the
forward probability calculation.
}
\examples{
# Ex1 (from tetrads)
set.seed(1234567) # For reproducability
l <- 10
rec <- rep(0.01, l-1)
n.tetrads <- 500 # number of spores to simulate
res <- sim_tetrad(n.tetrads=n.tetrads, l=l, rec=rec, p.assign=0.999,
   mu.rate=0, f.cross=0.8, f.convert=0.9, length.conversion=2, coverage=2.5)
snp.dat <- tetrad_to_df(res)
ddply(snp.dat, .(Tetrad, Spore, Chr), function(x){
    est_fwd_back(snp.dat=x, p.assign=0.999, p.trans=0.01)
    })
#
# Ex 2 (from single spores)
set.seed(1234567) # For reproducability
# simulate a recombination hotspot between the 100th and 101st snp
rec <- c(rep(0.001, 99), 0.1, rep(0.001, 99))
n.spores <- 500 # number of spores to simulate
spores <- sim_en_masse(n.spores=n.spores, l=200, rec=rec,
p.assign=.999, mu.rate=0.001, f.cross=0.5,
    f.convert=0.5, length.conversion=10, coverage=1)
snp.dat <- en_masse_to_df(spores)
ddply(snp.dat, .(Tetrad, Spore, Chr), function(x){
    est_fwd_back(snp.dat=x, p.assign=0.999, p.trans=0.01)
    })
}
\author{
Tyler D. Hether
}
\references{
Hohenlohe, P.A., S. Bassham, M. Currey, and W.A. Cresko. 2012. Extensive linkage
disequilibrium and parallel adaptive divergence across threespine stickleback genomes.
Phil. Trans. R. Soc. B, 395-408, doi: 10.1098/rstb.2011.0245

Drubin, R. S. Eddy, A. Krogh, and G. Mitchison. 1998. Biological Sequence Analysis: Probabilistic Models
of proteins and nucleic acids. Cambridge University Press, Cambridge CB2 8RU, UK.
}
\seealso{
\code{\link{simulate_coverage}}
}

