% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infer-recom.R
\name{id_recombination_events}
\alias{id_recombination_events}
\title{Identify crossover points from state sequences and snp locations}
\usage{
id_recombination_events(snps_genotypes_df)
}
\arguments{
\item{\code{snps_genotypes_df}}{a data.frame with two columns:
  \itemize{
     \item{snps}{ a vector of ordered snp locations (in bps)}
     \item{states}{ a vector of corresponding inferred states, either haploid
      (2 states = 0 or 1) or diploid with three states possible (0,1,2)}
   }}
}
\value{
a data.frame containing the midpoint between focal snp \eqn{i} and snp
\eqn{i-1} and whether their states were the same (0) or different (1).
}
\description{
This is a simple function that takes a vector of parental states
along a chromosome and identifies where states have changed.
}
\examples{
df <- data.frame(snps=100*(1:10),
   states=c(rep(0,4), rep(1,6)))
id_recombination_events(df)
}
\author{
Tyler D. Hether
}
\seealso{
\code{\link{fb_haploid}}, \code{\link{fb_diploid}}
}

