% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_parents}
\alias{make_parents}
\title{Simulate haploid parental genotypes along a chromosome}
\usage{
make_parents(snps)
}
\arguments{
\item{\code{snps}}{a vector gving the locations of snps along a chromosome}
}
\value{
an object of class \code{parent.genomes} giving the parental ancestry at each snp.
}
\description{
This function creates divergent genotypes between two parents. The first parent's genotypes
are denoted by a vector of zeros of length \code{L} while the second parent's genotypes are a vector of ones.
}
\examples{
make_parents(snps=c(1:10)*10)
}
\author{
Tyler D. Hether
}
\references{
none.
}
\seealso{
\code{\link{recombine_index}}, \code{\link{recobmine}}
}

