% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build-data.R
\name{make_snp_data}
\alias{make_snp_data}
\title{Function for building data}
\usage{
make_snp_data(parent_1, parent_2, snp_id)
}
\arguments{
\item{snp_spore_1}{A list of four elements, each containing a named vector of SNPs, indicating number of reads for each snp}
}
\description{
Function for building data
}

