% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{recombine_index}
\alias{recombine_index}
\title{Simulate recombination points along a chromosome.}
\usage{
recombine_index(scale, snps)
}
\arguments{
\item{\code{scale}}{either vector of length 1 specifying the genome-wide recombination rate (Morgans/bp)
or a vector of length(snps)-1 specifying the recombination rate between all snps. Rates must be positive.}

\item{\code{snps}}{a vector gving the locations of snps along a chromosome}
}
\value{
a vector of length \code{snps-1} specifying the location of recombiantion points.
}
\description{
Simulate recombination events given the vector of snp locations and recombination rate(s).
Haldane's function is used to simulate the probabilty of getting an odd number of crossover events between
any two snps given their physical distance and the specified recombination rate.
}
\examples{
set.seed(1234567)
# Simulating recombination across 1000 loci randomly spaced
# along a 200kbp chromosome
chromSize <- 2e5 # Chromosome size
l <- 1e3 # number of loci to simulate
c <- 1e-04 # Morgans/bp
snps <- sort(sample(size=l, 1:chromSize, replace=FALSE))
# Find recombination points between all the snps
recomb_points <- recombine_index(scale=c, snps=snps)
recomb_points
}
\author{
Tyler D. Hether
}
\seealso{
\code{\link{make_parents}}
}

