% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_en_masse}
\alias{sim_en_masse}
\title{Simulate random spores en masse}
\usage{
sim_en_masse(n.spores, l, rec, p.assign, mu.rate, f.cross, f.convert,
  length.conversion, coverage)
}
\arguments{
\item{n.spores}{An integer specifying the number of spores to simulate.}

\item{l}{an integer describing the number of loci to simulate.}

\item{rec}{a vector of length \code{l-1} that contains the the recombation rate
between each snp.}

\item{p.assign}{a numeric between 0 and 1 (inclusive) that gives the probability of
correct sequencing assignment.}

\item{mu.rate}{a numeric between 0 and 1 (inclusive) specifying the per snp mutation rate.}

\item{f.cross}{a numeric between 0 and 1 (inclusive) giving the frequency of recombination
events that result in crossing over. This is same as 1 minus the frequenc of non-crossovers.}

\item{f.convert}{a numeric between 0 and 1 (inclusive) that gives the frequency of gene conversion
during recombination.}

\item{length.conversion}{an integer specifying the mean (and variance) of a given gene conversion
tract.}

\item{coverage}{the mean (and variance) of sequencing coverage to be simlated.  \code{coverage} is
sampled from a poisson distribution (i.e., lambda=\code{coverage})}
}
\value{
A list of length n.spores. Each element of the list is of class \code{single.spore},
which contains three elements:
\describe{
	\item{p0.assign}{The number of reads that were simulated for parent 0.}
	\item{p1.assign}{The number of reads taht were simulated for parent 1.}
	\item{snps}{The snp id along the simulated chromosome.}
}
}
\description{
This is a wrapper function of many other functions that simulates a given number of
haploids each recombinant between two parents.
}
\examples{
set.seed(1234567) # For reproducability
# simulate a recombination hotspot between the 100th and 101st snp
rec <- c(rep(0.001, 99), 0.1, rep(0.001, 99))
n.spores <- 500 # number of spores to simulate
spores <- sim_en_masse(n.spores=n.spores, l=200, rec=rec,
p.assign=.999, mu.rate=0.001, f.cross=0.5,
    f.convert=0.5, length.conversion=10, coverage=1)
}
\author{
Tyler D. Hether
}
\seealso{
\code{\link{recombine}}, \code{\link{make_parents}},
\code{\link{simulate_coverage}}, \code{\link{recombine_index}}, \code{\link{id_hotspots}}
}

