% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_tetrad}
\alias{sim_tetrad}
\title{Simulate recombination of a given number of tetrads}
\usage{
sim_tetrad(n.tetrads, scale, snps, p.assign, mu.rate, f.cross, f.convert,
  length.conversion, coverage)
}
\arguments{
\item{n.tetrads}{An integer specifying the number of tetrads to simulate.}

\item{p.assign}{a numeric between 0 and 1 (inclusive) that gives the probability of
correct sequencing assignment.}

\item{mu.rate}{a numeric between 0 and 1 (inclusive) specifying the per snp mutation rate.}

\item{f.cross}{a numeric between 0 and 1 (inclusive) giving the frequency of recombination
events that result in crossing over. This is same as 1 minus the frequenc of non-crossovers.}

\item{f.convert}{a numeric between 0 and 1 (inclusive) that gives the frequency of gene conversion
during recombination.}

\item{length.conversion}{an integer specifying the mean (and variance) of a given gene conversion
tract.}

\item{coverage}{the mean (and variance) of sequencing coverage to be simlated.  \code{coverage} is
sampled from a poisson distribution (i.e., lambda=\code{coverage})}

\item{l}{an integer describing the number of loci to simulate.}

\item{rec}{a vector of length \code{l-1} that contains the the recombation rate
between each snp.}
}
\value{
A list of length n.tetrads of class \code{tetrad}. Each element of
\code{tetrad} is itself a list of class \code{individual.tetrad}, which
contains four lists, one for each spore, and three elements:
\describe{
	\item{p0.assign}{The number of reads that were simulated for parent 0.}
	\item{p1.assign}{The number of reads taht were simulated for parent 1.}
	\item{snps}{The snp id along the simulated chromosome.}
}
}
\description{
This is a wrapper function of many other functions that simulates a given number of
tetrads each with four haploid spores that are recombinants between two parents.
}
\examples{
set.seed(1234567) # For reproducability
# simulate a recombination hotspot between the 99th and 100th snp
l=200; scale = 0.01; snps=2*(1:l);
n.tetrads <- 100 # number of spores to simulate
res <- sim_tetrad(n.tetrads=n.tetrads, l=l, scale=scale, snps=snps, p.assign=0.999,
   mu.rate=0, f.cross=0.8, f.convert=0.9, length.conversion=10, coverage=2.5)
}
\author{
Tyler D. Hether
}
\seealso{
\code{\link{recombine}}, \code{\link{make_parents}},
\code{\link{simulate_coverage}}, \code{\link{recombine_index}}, \code{\link{id_hotspots}}
}

