% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_coverage}
\alias{simulate_coverage}
\title{Simulate sequencing across simulated data:}
\usage{
simulate_coverage(simdata, p.assign, coverage)
}
\arguments{
\item{simdata}{an object of class \code{recombine}}

\item{p.assign}{a numeric between 0 and 1 (inclusive) that gives the probability of
correct sequencing assignment.}

\item{coverage}{the mean (and variance) of sequencing coverage to be simlated.  \code{coverage} is
sampled from a poisson distribution (i.e., lambda=\code{coverage})}
}
\value{
an object of class \code{snp.recom} providing the simulated sequencing reads for each parent.
}
\description{
This function simulates sequecing across simulated data of class \code{recombine}.
}
\examples{
set.seed(1234567) # For reproducability
l <- 1000 # number of loci to simulate
rec <- 0.01 # recombination rate between each snp
r <- recombine_index(rep(rec, l-1)) # recombination rate between each snp (vector form)
p_a <- .999 # probability of correct sequencing assignment (1-sequence error rate)
p <- make_parents(l) # make the parent
recomb_sim <- recombine(parents=p, r.index=r, mu.rate=0, f.cross=.5, f.convert=1, length.conversion=10) # recombine parents
sim_reads <- simulate_coverage(simdata=recomb_sim, p.assign=p_a, coverage=1) # simulate sequencing coverage
}
\author{
Tyler D. Hether
}
\seealso{
\code{\link{recombine}}, \code{\link{estimate_anc_fwd_back}}
}

