% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forward-backward.R
\name{tetrad_to_df}
\alias{tetrad_to_df}
\title{Convert from class tetrad to class data.frame}
\usage{
tetrad_to_df(class.tetrad)
}
\arguments{
\item{class.tetrad}{a list of class tetrad}
}
\value{
a data frame with 6 columns, c("Tetrad", "Spore", "Chr", "Snp", "p0", "p1")
}
\description{
Converts from simlated data of class tetrad to class data.frame
}

