% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_parents}
\alias{make_parents}
\title{Simulate haploid parental genotypes along a chromosome}
\usage{
make_parents(snps)
}
\arguments{
\item{L}{is an integer specifying the number of loci to simulate. L must be greater than 1.}
}
\value{
an object of class \code{parent.genomes} giving the parental snp calls at each of the \code{L} loci.
}
\description{
This function creates divergent genotypes between two parents. The first parent's genotypes
are denoted by a vector of zeros of length \code{L} while the second parent's genotypes are a vector of ones.
}
\examples{
# Make a chromosome of L=50 loci for each parent.
make_parents(L=50)
}
\author{
Tyler D. Hether
}
\references{
none.
}
\seealso{
\code{\link{recombine_index}}, \code{\link{recobmine}}
}

