% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{recombine_index}
\alias{recombine_index}
\title{Simulate recombination points along a chromosome.}
\usage{
recombine_index(scale, snps)
}
\arguments{
\item{p.trans}{a vector of length \code{L-1} specifying the recombination
rate between two snps. Each element in \code{p.trans} must >= 0 (no recombination)
and <= 0.5 (independent assortment).}
}
\value{
a vector of length \code{L-1} specifying the location of recombiantion points (i.e., between snps).
}
\description{
Simulate recombination events given the vector transistion
probs and the number of loci.
}
\examples{
set.seed(1234567)
l <- 100000 # number of loci to simulate
scale <- 0.01 # d = bp*scale
snps <- 1:l # floor(seq(from=1, to=1000, length.out=l)) # position of snps (in bps)
# snps[25:l] <- 5000+snps[25:l]
sum(recombine_index(scale=scale, snps=snps))/l
}
\author{
Tyler D. Hether
}
\seealso{
\code{\link{make_parents}}, \code{\link{fitEcm}}
}

