% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_en_masse}
\alias{sim_en_masse}
\title{Simulate random spores en masse}
\usage{
sim_en_masse(n.spores, scale, snps, p.assign, mu.rate, f.cross, f.convert,
  length.conversion, coverage, chr.name = "I")
}
\arguments{
\item{n.spores}{An integer specifying the number of spores to simulate.}

\item{p.assign}{a numeric between 0 and 1 (inclusive) that gives the probability of
correct sequencing assignment. A value of 1 means that no sequencing error or ancestral
polymorphism is to be simulated.}

\item{mu.rate}{a numeric between 0 and 1 (inclusive) specifying the per snp mutation rate.}

\item{f.cross}{a numeric between 0 and 1 (inclusive) giving the frequency of recombination
events that result in crossing over. This is same as 1 minus the frequency of non-crossovers.}

\item{f.convert}{a numeric between 0 and 1 (inclusive) that gives the frequency of gene conversion
during recombination.}

\item{length.conversion}{an integer specifying the mean (and variance) of a given gene conversion
tract (in bps).}

\item{chr.name}{a numeric or character value specifying the chromosome name (default to "I")}

\item{\code{scale}}{either vector of length 1 specifying the genome-wide recombination rate (Morgans/bp)
or a vector of length \code{snps-1} specifying the recombination rate between all snps. In either
case rates must be positive.}

\item{\code{snps}}{a vector gving the locations of snps along a chromosome}

\item{r.index}{a vector of length \code{snps-1} specifying whether a recombination is to be
simulated (1) or not (0) in between two adjacent snps.}
}
\value{
A data.frame of class \code{en.masse}. Each row contains the following information:
which contains three elements:
\itemize{
	\item{Spore}{ The spore ID (1:length(n.spores))}
	\item{Chr}{ The chromosome name}
 \item{Snp}{ The snp location (in bps)}
 \item{p0}{ The number of reads that mapped to parent 0}
 \item{p1}{ The number of reads that mapped to parent 1}
 \item{states_given}{ The simulated states (0 or 1)}
}
}
\description{
This is a wrapper function of many other functions that simulates a given number of
haploids each recombinant between two parents. As the name implies, information of the all four
products are lost since spores are random. See \code{\link{sim_tetrad}} if all four haploid recombinants
are desired.
}
\examples{
# Simulating 5 haploid recombinants
set.seed(1234567)        # For reproducibility
n_spores <- 5          # number of tetrads (meiosis events)
l <- 75                  # number of snps to simulate
c <- 3.5e-05             # recombination rate between snps (Morgan/bp)
snps <- c(1:l)*1.3e4     # snps are evenly spaced 20kbp apart
p_a <- 0.95              # assignment probability
coverage <- 2.1          # mean coverage
# Now simulate
sim5 <- sim_en_masse(n.spores=n_spores, scale=c, snps=snps,
	p.assign=p_a, mu.rate=0, f.cross=0.8, f.convert=0.3,
	length.conversion=2e3, coverage=coverage)
sim5
}
\author{
Tyler D. Hether
}
\seealso{
\code{\link{recombine}}, \code{\link{make_parents}},
\code{\link{simulate_coverage}}, \code{\link{recombine_index}}, \code{\link{id_hotspots}}
}

